Built The Seam System
Added a rail-based seam system that allows the player to use the seam-ripper and needle to move a point along a prescribed track on the bear.

Saved the track as a prefab, so that more instances can be built. More nodes can be added manually in-editor by adding new node objects to the array

Known Bug: when the mover object reaches either end of the seam, it throws an exception because it is exceeding the array bounds. This only happens once each time it reaches the ends.