Implemented Blake's prototype Environment, implemented New Music
Put the second environment in a new scene separate from the other stuff; imported all of the sound effects, and implemented the background music. Changed the clean material so the seam could be more easily read.

I tried to implement the drawer opening system to Cody's drawers but that didn't work, not sure why